requirements.txt
alembic==1.5.4
canonicalwebteam.flask-base==0.8.0
canonicalwebteam.http==1.0.3
canonicalwebteam.blog==6.3.1
canonicalwebteam.search==1.0.0
canonicalwebteam.templatefinder==1.0.0
canonicalwebteam.image-template==1.3.1
canonicalwebteam.discourse==3.0.8
canonicalwebteam.launchpad==0.8.2
python-dateutil==2.8.1
pytz==2020.5
canonicalwebteam.store-api==3.1.7
maxminddb-geolite2==2018.703
Flask-OpenID-Stateless==1.2.6
feedgen==0.9.0
feedparser==6.0.2
pymacaroons==0.13.0
pybreaker==0.6.0
Flask-WTF==0.14.3
requests==2.25.1
sqlalchemy==1.3.23
marshmallow==3.10.0
mistune==0.8.4
psycopg2-binary==2.8.6
pyyaml==5.4.1
ubuntu-release-info==21.1
launchpadlib==1.10.13
macaroonbakery==1.3.1
sortedcontainers==2.3.0
vcrpy-unittest==0.1.7
pillow
